improved QueryFetcher for suspends fetch
getOrCreateQueryFetcher function created. Allows to create a new QueryFetcher or retrieve it from the cache. In case of useQuery it is always created.

changed disposeRetain: now also delete the previous query from the cache

changed isDiffEnvQuery: now the query is considered different if it has a different identifier

changed fetch: added suspense parameter, when asked to suspend the fetch the promise is throw and the forceUpdate blocked