Improve useSubscription hook types useSubscription allows to omit `config.subscription` and `config.variables` if `opts.skip` is literally true. I.e. `subscription` and `variables` are optional only when we explicitly pass true to skip option: `useSubscription(config, { skip: true })`.
But in reality we never do it like this `skip: true`. We always use some variable like this for example `useSubscription(config, { skip: !userId })`. But in this case with current typescript if we use `skip: !userId` subscription and variables are required.

This change makes `config.subscription` and `config.variables` whenever we provide skip option. The value of skip option can be known in runtime only